drivers: mathworks: fix sprintf argument warning
In function mw_stream_chan_show the warning is:
'''
drivers/misc/mathworks/mw_stream_channel.c:1050:25: error:
format ‘%llu’ expects argument of type ‘long long unsigned int’,
but argument 3 has type ‘long int’ [-Werror=format=]
'''

(cherry picked from commit 69a8743b524d71ff71c272b1373500e2240ac1ea)

+++ drivers/misc/mathworks/mw_stream_channel.c
@@ -1047,7 +1047,8 @@ static ssize_t mw_stream_chan_store(struct device *dev, struct device_attribute 
 static ssize_t mw_stream_chan_show(struct device *dev, struct device_attribute *attr,
         char *buf)
 {
-    return sprintf(buf, "%llu\n", atomic64_read(&rxcount)); 
+	return sprintf(buf, "%llu\n",
+		       (unsigned long long)atomic64_read(&rxcount));
 }
 
 static DEVICE_ATTR(dma_irq, S_IRUGO, mw_stream_chan_show, mw_stream_chan_store);
